A few fixes to get build-gov-aws-amazonlinux builds completing
In #383, the yum-managed pip is upgraded with pip. As part of that
upgrade pip removes /usr/bin/pip and install /usr/local/bin/pip. This
will likely break behavior of any root processes trying to use pip or
even yum updates of the pip package. That being said, we update our prep
script to resolve the new location so that it can get salt ready to go.